+++ lib/feature/transactions/viewmodel/transactions_viewmodel.dart
@@ -5,6 +5,7 @@ import 'package:browny_applications_new/core/data/remote/models/api_model_index.
 import 'package:browny_applications_new/core/data/remote/models/content_localize_data.dart';
 import 'package:browny_applications_new/core/data/remote/models/request/payment_check.dart';
 import 'package:browny_applications_new/core/utils/crashlytics_helper.dart';
+import 'package:browny_applications_new/core/utils/repo_result.dart';
 import 'package:browny_applications_new/feature/authentication/error/authen_exception.dart';
 import 'package:browny_applications_new/feature/transactions/models/machine_program_model.dart';
 import 'package:browny_applications_new/feature/transactions/models/payment_transaction_state.dart';